+++ types/ember-cli-mirage/index.d.ts
@@ -42,7 +42,7 @@ interface ModelInstanceShared<T> {
     toString(): string;
 }
 
-export type ModelInstance<T> = ModelInstanceShared<T> & Model<T>;
+export type ModelInstance<T = AnyAttrs> = ModelInstanceShared<T> & Model<T>;
 
 interface Collection<T> {
     models: Array<ModelInstance<T>>;
@@ -208,4 +208,6 @@ export class JSONAPISerializer {
     keyForAttribute(attr: string): string;
 
     keyForRelationship(relationship: string): string;
+
+    serialize(object: ModelInstance, request: Request): Document;
 }